Separate `benchmark_accelBLAS`
