+++ src/Platforms/tests/CMakeLists.txt
@@ -32,3 +32,17 @@ if(USE_OBJECT_TARGET)
 endif()
 
 add_unit_test(${UTEST_NAME} 1 1 $<TARGET_FILE:${UTEST_EXE}>)
+
+if(BUILD_MICRO_BENCHMARKS)
+  set(UTEST_EXE benchmark_accelBLAS)
+  set(UTEST_NAME deterministic-unit_${UTEST_EXE})
+
+  add_executable(${UTEST_EXE} benchmark_AccelBLAS.cpp)
+  target_link_libraries(${UTEST_EXE} platform_LA platform_runtime catch_main containers)
+
+  if(USE_OBJECT_TARGET)
+    target_link_libraries(${UTEST_EXE} platform_omptarget_LA)
+  endif()
+
+  add_unit_test(${UTEST_NAME} 1 1 $<TARGET_FILE:${UTEST_EXE}>)
+endif(BUILD_MICRO_BENCHMARKS)
